Updated the multi site directoy resolving
Moves the all of the files that need to be edited for multiple sites to work to inside of the root/sites directory.  This is good because now commits don't need to have all of the sites we're working on sent out to the whole world, and we think it will be less confusing for multiple sites to work as well.