Mainly an upgrade to Permissions plugin, but also includes some clean up.
This was a pretty major upgrade to the Permissions plugin. It is now much easier to update and maintain available access control methods. Instead of having to manually update and manage available actions, you simply click a button, and when a new plugin or controller is added you only need enter the name, and all sub actions will automatically be found and added as well.  This second part also stems from an upgrade to the app_controller build_acl method, which now allows you to build_acl for a single plugin or controller, instead of being forced to rebuild the entire system every time.  Unfortunately, the commit also includes some minor clean up as well due to a lack of timely commits, so there may be some unrelated errors (to permissions) that were caused here, because of deleting and cleaning unecessary files and methods.