Had to move app_controller, and app_model back
There was an unforeseen problem caused by moving app_controller, and app_model and it was too cumbersome to make a fix, when there wasn't an absolutely great reason for these files being moved in the first place. So they were put back, and now /sites directories now work again.