Created a database upgrade method in /admin/index
It is now possible to upgrade the database and the current version number with a single click.  When you visit /admin a check is run to see if a later version number db file exists in /version directory, if it does it shows you the sql to run, and a button to run it. In order for this to work, you must add a new system setting though, ZUHA_DB_VERSION:0.0001.sql into the SYS key settings.