Added some todos and updated build_acl to work with multi-sites
This is good because now you can run build_acl and expect it to work even if you are working from a site not using only the core app directory. Left a few todos related to this, but mostly verification todos regarding items I'm not 100% certain are working for every use case I can think of.