Updated catalogs plugin to get rid of the prices table
We have a standard, which calls for us to name plugin tables according to the plugin name, so in this case the plugin is catalogs so all tables for this plugin should start with catalogs. Therefore we dumped prices, and started using catalog_item_prices, which consequently already exists.  Please note, not all of the fields in catalog_item_prices are currently being used, and that will be a future update.