Upgraded attributes, and added attributes to con people adding
This upgrade now makes use of a reusable element and reusable functions in the model.  It marks a shift to fat model, and skinny controller coding, and much better use of reusable code overall. It also make the creation of custom form fields for use in various system database tables much easier to add.  Though we will continue to simplify it even further.