Clean up for Attributes
Added some better linking, and rudimentary filtering so that you could find your way around attributes better.  This is in anticipation of making a new feature for system attributes which mirror the fields already existing in database tables, for the purpose of having database control over the form display on system pages.