Updated some more linking and filtering in attributes
Just making it so that you can poke around using the is_system and group as a limiter for index return results.  This is necessary because once we add a lot of mirrors of the currently existing default attributes, we will have tons of attributes, far too many to display without limiters.