You can now load system forms from the webpages wysiwyg editor.
With this upgrade a small snippet of text (ex. {form: Plugin.Model.type.limiter}, like this {form: Contacts.ContactPerson.add.59}) allows you to include any form you've managed with through attributes and attribute groups to be loaded. Attribute Groups = forms, Attributes = form inputs.  Plugin = Camelcased Plugin name, Model = Camelcased model name, type = add, edit or view, limiter = is used to denote a particular grouping for attribute_groups.  But note... if the value in the db for enumeration_id is null, it is pulled as well if it matches the model.