Major upgrade to forms plugin.
Lots of bugs and now the forms plugin processes all of the forms. This change was made so that we could have redirect urls specified without going through every controller and adding a check for successUrl or failUrl.  The creation, and submitting of user controlled forms has been completely tested.  OH, and almost forgot, also updated the forms element so that it takes advantage of all of the form settings available in cakephp.