Zuha Just Became an APPLICATION BUILDER
With the work done today we have just setup the basis for a complete application building program. Because the user can create forms and control which fields are addable, editable, and viewable, we've basically moved the entire database management to the user end, so that they can control the data each table uses. Yesssssss