The beginning of a workflow and conditions overhaul.
This is the conditions side of the update, and is good because we have simplified the conditions table, and rethought the condition process to be more compartmentalized.  Now instead of controlling part of the save, we have restructured so that you would simply do an on the fly hasMany bind of the conditions table to the model - probably a plugin - that belongsTo the condition. This allows you to put all of the post condition meeting criteria into the model which wanted the condition to be met. ie. compartmentalized.