Another update to how conditions work.
Seems like it should have been this way all along. But now, we will no longer need to pass this-> params to all models, because now we will check afterSave using just create, and update types, while saving delete and view condition types for a later upgrade.  Its so much more elegant like this, but now we will be checking for conditions on all create, read, update and delete procedures to see if a condition was matched.  That may be a big of a lag on performance, but with indexing it should be relatively quick, and because they're compartmentalized into crud types its much clearer whats going on.