Conditions now fire the an action called 'triggered' from the binded_model field of the condition.
This is a pretty substantial change to the conditions, in that the only thing that conditions do is fire a single predetermined action which is specified by the bind_model field. As previously mentioned, it takes all of the saving out of the conditions and moves it to the binded model.  So now you can have the 'triggered' action do anything you'd like it to do.  This is good because its hugely more extensible now, because any model could have a triggered action, and to fire that anytime, you would just create a condition.  That is pretty cool.