Workflows plugin has been baked.
This new plugin will make it possible so that a single action performed on your website can trigger multiple other actions to happen.  The structure is also setup to allow a tree of events, so that you can have certain items lead to other items, based on the previous event taking place. It moves zuha much closer to a complete application builder.