A substantial update to profiles registration
We now do a lot of case checking when registering to see if a contact or a user exists or does not exist, and then during registration create that contact or create that user.  This is good because now you can make users out of contacts, and make profiles for existing users - if they don't already have one. Also did a lot of checking on these different cases (8 total) and the validation critieria for creating and logging in as users.  The last major upgrade needed for users, is an email validation being required or not.