All ckeditor calls have been updated.
This change updates all places where ckeditor was being loaded, so that they now load using the richtext, type option, and buttons option that was provided with an earlier update.  At this time we also updated the richtext form input type so that you could specify the buttons available to the editor.