Do Not Use This Commit Except for a Fallback
There is some good code in this commit in the app_controller and in the app_model, and in the user model which I'm committing so that we have a record of it if needed.  It has to do with ACL, and page access.  Much of it will be deleted after this commit, because I believe we have found it to be more efficient to use straight queries.  -- we have saved one second per page load at least, and would be way more if we cache that query -- Ultimately the hope is that this makes our system more scalable over the long term, because even with barely any records, acl was pulling 7 queries at 1-2 seconds per page.