Acl is in the books.
The only items left are, 1. check for negatives, 2. make it so that a aros_acos record is created when a record level page is added, 3. fix the sql to update settings.   Other than those three, we have a fully functional record level access control system.  Which is way faster than it was before, thanks to the custom queries.  Acl component is not even necessary anymore, thanks to this update.