Acl Record Level Access Updated.
When you userLevel is set in the controller to true, a record level aco is created (as it was), but now the aro with access is created in the table aros_acos.  Meaning record level access on the fly.  This only works when you access records by id though, no named parameters, so a new best practice is to redirect to the id of the record by using beforeFilter in the controller for your redirect cases. It has many benefits besides this, because it also makes your views and edits much easier to review.  The next step is to make it so that we create field level access, for things like assignee_id, user_id, creator_id, etc.