Major code deletion of userLevel and userField from all models.
Previously these variables were used to make a model available for record level access control.  And it is no longer needed because we have implemented database driven settings for this.  This is WAY better for our multi-site setups because now there is no hard coding of what is an app level setting.