Acl Change Again (Back to Using Acl Component).
It was realized that acl checks were faster if we used them to their full potential.  So guests and record level Acl is really getting narrowed down to many fewer lines of code.  Deleted all of the access check in the app_model, and got rid of the record level aco check because we could simply check if the aco existed as a record level instead, and check it by record if it does.  Much nicer piece of access code.  IMPORTANT NOTE : This made it so that we cannot use negative values in our Aros_Acos table.  But we deny by default, so we merely have to make sure that if we do grant specific permissions to a record, then we make sure to define every aro that has access for that record.  Not just the user who does.