Added Profile Image and Link Element, and Updated Default Layout File.
The element just gives an easy way to display your profile image, and links to edit or view your profile.  However the bigger change with this, is that we fixed the naming of database driving template tags.  We changed {element} to helper, and made {element} actually work with elements instead of helpers. This is good because the naming makes more sense, and because it will make it easier to create elements for the site without needed to edit app_controller and include the helper in the helper call.