Fixed a bug in the snpsht element, added backtrace function to the core, and made cache of elements user specific.
This is kind of a temporary fix to the problem of creating user specific elements (like a profile avatar) and caching them for speed purposes.  Also added a little documentation, and fixed a bug which was a real pain.  When doing requestAction from elements, it can get very tricky to find out why ACL is not working.