Updated View Path Find Function.
This change makes the view path find function a bit lighter, and a tad easier to follow by moving everything into an array, and checking each file in the array to see if it exists, and after finding the first file that exists it returns the viewPath which matches.  This makes the layering of view files and localized view files, and multi-site view files a bit clearer.