Fully Supported Extension Parsing.
All sites now have the ability to parse any extension that there is a view file for.  (Including a default view file, so that you don't have to have a view file in every single place that you want to use a standardized view.)  Currently support for json view, and index, and xml view are done.