Scaffolded Affiliates and Credits Plugin.
These two plugins will work hand in hand but separate enough to be two different plugins.  Affiliates will be responsible for tracking user sign up referrals (mainly).  Credits will be used for giving users some kind of value to the actions taken on the site.  I do suggest that we mainly, create functions and use Workflows and Conditions for the use of these Model functions.