Changed Gallery Display to Elements.
Previously galleries were displayed using a helper.  It has been switched to using elements, where the name of the element is the type of gallery you want to display.  This is good, because if you want to add new gallery types, you would add them to the gallery type array, and add a gallery element which uses the fields.  Its a nice upgrade, but I'm sure we could take it a bit further and make the types database driven as well.