Updated Profile snpsht element, and Created Gallery thumb element.
The profile element now calls the gallery element if you use the settings, useGallery, and send the profileId, otherwise it uses the soon to be dprecated avatar_url html.  The gallery element calls the thumb image and links to the gallery by default. That can be over-ridden with the plugin array.  Overall this is a great update, because it makes use of an easy way to display the users profile thumb, and the gallery thumb anywhere on the site, with a simple element call to the 'thumb' element.