Major Change - Users Moved to Plugin, Cupcake Forum Deleted, and new Message Plugin.
By combining users and profiles into a users plugin we will make two things which are 100% related much easier to find and manage.  The cupcake forum was deleted because it does not make integration into an existing user system easy, and so its easier to just write our own forum plugin. And the Message plugin was moved from profiles to its own plugin, for more extensibility.