Had to Remove Caching for Form Elements.
There does not seem to be a way of having form inputs control whether the form is cached, so the form element cannot be cached.  This will be a BIG PROBLEM, for any site that wants a created form on every page of the site, or if they wanted multiple forms on every page of the site.  It would be a major problem. Not sure how we fix it, except adding it to element call with something like, {form: id/add.cache}