The beginning of a new way to handle settings.
This new method of handling settings is VASTLY superior.  It is not finished, but the plan is to make it file based, and the file will be written each time the settings table is updated.  Now we can have separate records for each and every setting, and by using serialize and unserialize it will be available in associative arrays throughout the application.