Forms Upgrade.  You can now fully utilize edit and view types of the form.
This upgrade made it so that you can use template tags {form: 1/edit} and {form: 1/view} to show is_editable and is_viewable values for custom forms.  The big limitation is that you must use them through the context element first, because you must show these form in a context where the data is already available.