Got rid of the Ajax, and Javascript Helpers.
Cakephp has deprecated Javascript helper, and we made very little use of it. It is now gone, and the Js helper has been put in instead for the small number of instances where needed.  The ajax helper is probably not worth using either because the Js helper should probably cover anything it can do.