Fixed Bug in Categories Config, Added New Variable to Scaffolds.
This update makes it so that you can set a pageHeading in your settings file and have it reflected on scaffold pages using the pageHeading variable name.  Also fixed a bug in categories_app_controller, which could possibly need more work, as I'm not sure why the bug was there in the first place.