System Javascript House Added.
A folder in the system webroot has been added to house system javascript that might be needed on system pages.  It should be used for edit pages, etc, which need javascript that other pages don't. And it should contain only necessary and widely used system javascript.  Also, it should be called using a script call from the view file for the page it is needed for.