Public Side Css Editing.
You can now edit css files via the public side of the site, and the css files are used for display using the default layout.  This update was made so that the CMS could handle both html, and css in a database driven way.  We'll probably do this for javascript as well, though there are security issues there, which might prevent us from doing so.