Added Support for Stylesheets in CKEditor.
You can now set the contentsCss option in the richtext form input type, and you will have support for showing the content in the editor with a style sheet associated with it.  This was done so that pages of the template type could be a bit more visually appealing, while editing.