Removal of Excess Views : Wiki Page View.
This update removes the view.ctp file and lets wikis use plugin configuration to control the display of wiki pages.  Its of important note, that with this update the Scaffold view now supports contained data.  So that you can put related data into views, using Model.field format in the defaults or settings dot ini files.  This is good because it continues the effort to reduce the number of view files required.