Added Recursive Data to Conditions.
So that we could do related model data lookups when firing workflows, we added recursive data to the serialized data record that gets saved when a condition is met. This is good because it adds strength to the workflow abilities, but might be a performance hit because of the recursive lookup that has to be performed after you've already done the work of looking up if the condition was met.