Affiliate Codes Verified to Work.
If your registration form includes the field [referral_key], and your model actsAs = Affiliates.Referrable, then the affiliated table will be updated to let you know who referred that record save.  The best part about this is that any record can be referred, very easily now, and though it is only turned on for the User model, it takes the addition of a single form field, and the actsas variable in the model to start referring other records.  To test this, add a field called with the name referral_key to the User registration form, and add a referral code from the affiliates table when registering.  You should then see a new record in the affiliated table after registration.