Updated All config/bootstrap.php Files.
This update makes it so that there is a single file required by all bootstrap files to store global functions which are a level above the global functions in bootstrap.php.  This is a good update because it will be much more rare that there will be a need to update 10's or hundreds of bootstrap.php files.