Recoded OrderTransaction add()
This was needed because saveAll and updateAll do not fire callbacks (like afterSave and beforeSave), so we have to implement a new policy of not using the All functions, because it interferes with our ability to use conditions and workflows effectively.  With this update you can see OrderTransaction::add() and see how model save functions should work.  They save each item within each model, and then do a roll back if it was unsuccessful at any point.  If anyone has anargument for using saveAll or updateAll in light of the callbacks not being used, I'm up for debate about it so that the right solution is met.