CatalogItem Index By User Role and Price Now
The catalog item index shows the right price, and/or doesn't show the item at all by filtering with user role, and price settings for user role. If no advanced price matrix exists it shows the default price to all users.  If advanced price matrix does exist then we do not show the item if the price is empty for the user role of the current user, otherwise we show the price specific to the viewers user role.

This update is good because now we have moved closer to having unlimited pricing differences for a single item.  I see this as the future of being able to test multiple prices, and of course for whole sale vs. retail pricing.

Along with this update some convenience functions were added as well.  Including enum() in the global function file at /app/config/global.php, which allows you to very easily find enumerations table record ids.  And ->userRoleId was added to the app_controller so that all controllers have access to that value easily.  And that value is set to guest user role id if it is empty.