Guest Cart, and User Role Pricing.
This update has fixed as many errors as I could find with the guest cart, and user role pricing.  It seems to be working, but I'm going to do some more testing here shortly.  But now you can at least be a guest, and add items to a cart, and then when you go to checkout you will be asked to login, and then your cart is merged automatically when you go back to the checkout page.