Streamlined Admin UI Database Updates.
With this update you no longer need to refresh the admin to see if there is a nother database query to run.  You just click, upgrade, upgrade, upgrade until the button disappears.  A step was removed, and therefore this is good because its faster to update the db now by visiting /admin of any site.