Added Favorites Plugin, and Customized.
Added the favorites plugin from CakeDC, and customized it to be a bit easier and work better with the zuha system.  One specific goal is to get rid of the types declaration because it is too limiting, allowing only a single model per type, (ie. favorite = CatalogItem, means you can't have favorite Category or something like that, unless you make a new name, like favoriteCategory.)  That might be nice, if you need to separate, but it is just too limiting in my opinion and there are other ways to separate.