Added User Role Functionality to Context Element.
Before this update you could only set contextual pages using the url someone was looking at.  You can now set contextual elements to appear based on the userRole that is looking at the page.  This is good because it is the base of allowing the end user to control views by userRole of the viewer.  To test setup a template tag like this {element: webpages.context.1} and in the settings table add __ELEMENT_WEBPAGES_CONTEXT_1  where the values are equal to userRoleId = 3 and pageId = 10.  You will then see page id 10 in place of that template tag, if the user role is 3 of the viewer looking.