Cleanup and Simplify New Site Setup.
In order to make new site setups easier I have created /sites/example.com to be used as a template for installing a new site to an existing setup that uses multi-sites.  I have also updated the zuha_dirty.sql, and zuha_clean.sql with the latest database schema.  And lastly I deleted a number of now unecessary files and db tables.