Created Webpage Js Functionality.
You can now upload js files via the user interface and they will automatically be added to the default template.  So when you are created a new site you no longer need to add files to a customized layout in order to get javascript links to work. To test this use the dashboard bar at the bottom when you're logged in as admin, and click [Webpage Js], then upload some javascript text.