Forum SQL Added.
The sql in 0.0082.sql will make it so that new installs will have the forum plugin working by default.  Previously you would have had to go in and manually make the admin user a forum admin.  Though, I'm pretty certain that there is a problem when new users register they will not automatically have their forum access level updated.  Not tested, but as soon as someone needs a forum we will test and make sure it works.