Updated Admin Navigation.
I still don't much like the order of everything, but needed to add a lot of links because there has been many additions since the last time this navigation was updated.  Way more functionality.  If anyone (ie. Amy) wants to look at the admin navigation and suggest a better organization of it, that would be cool.