Made an element allowed at all times.
There is a bug with the acl check when using site wide template tags for an element which using a function via requestAction.  Basically, because the function is not allowed explicitly in your permissions you get redirected to the login page.  But, the login page is also using that site wide template, and that site wide template tag.  So there is an infinite redirect loop going on which is difficult to figure out and debug.  I'm not quite sure how to fix it yet, so I put in a temporary fix which allows the action by hard coding it into allowedActions.  If any one has any thoughts on how to fix it, please share.