Page Editing Now Works for All Templates
Previous to this you could only edit global content blocks using the default template.  You can now edit global content blocks while viewing any page which is of the [template] type.

This is good, because now with Multi-Templates you still retain the ability to edit easily even when you're viewing a page not using the default site template.

You can test this typically by viewing the homepage, and trying to edit a global content box.  Most of the time the homepage is using a different template than the homepage template. But not always, FYI.