Added System CSS File to Default Template.
Added a file to /app/webroot/css/system.css and included it in the /layouts/default.ctp file.  This is good because now we can make system css which is available to all sites when they are installed, and now when you visit system pages they will at least have a default layout for common areas (like Blog Comments for example).  These default css rules can be easily over written by site css rules, and should be kept a minimal and as simple as possible.  They are ONLY meant so that when you view a system page it doesn't look like garbage because you haven't done the css for that section for that particular site.