Added Social.Share Plugin, Upgraded Helper Settings.
Added a simple social share plugin that uses the TagCallback helper by default to add facebook like, and tweet buttons just after the content_for_layout callback on anypage that the helper is loaded.  It needs to be configurable, but at this point you simply get the two buttons and you like or tweet the current page that you are on.  You can test it by adding, APP_LOAD_APP_HELPERS setting with the value Social.Share or helpers[] = Social.Share, or CamelCaseController = SocialShare and you will see the like buttons automatically appear every where or on the specific pages that you have set the settings for.

In doing this, the app helper settings were also updated to include multiple helpers per controller, where as before if you were including a helper dynamically on a per controller basis, you could only include a single helper per controller.