Attempting to finish an auto generated acl controlled menu.
This menu is supposed to show all of the pages you have access to (for the current logged in user) and thus would automatically show links to tickets, catalogs, etc if they are turned on in ACL for your user role or user id.  However its very slow, and I'm not sure we're going to keep it like this so it could be changed shortly or removed.