Sweet Update! Html-link() now Permissions Based.
Added an option to the html->link() function so that if you set checkPermissions => true, in the options array before showing the link it will check if the user has permission to view that page.  This makes the whole system much better because now pages which you don't have permission to will not appear, and instead of clicking a link and getting the No Permission message there will just not be a link to click.

Note, I have a todo on this for adding an alternativeText option, so that instead of showing the link we can show alternative text, like [Login to View X Page].

This will make contextual link sooooo much easier its amazing.  There are some performance issues though, and we need to be careful about having too many links that checkPermissions on a single page.

Further, I made this an option that you have to declare as true, instead of auto checking every link for a few reasons.  1. for performance, 2. for backwards compatibility, 3. because some links just don't need to have their permissions checked.