Renamed Forum Views and Controllers Due to Conflict.
There was a conflict in ACL which was making it impossible to set permissions for the users plugin, because forums also contained a users_controller.  So the solution was to add the forum_ prefix to all of the forum controllers and view folders, and to update all of the links.  There is still a naming conflict with the models, but so far I have seen no ill effects because of it, so if it ain't broke, don't fix it... on that one.