First pass of ajax form validation done.
The forms element now supports ajax validation of the fields you input for email, and number checking, and whether the field is required or not.  Additional validation checks can be added as needed using the form_inputs table.  This update is good because you get instant feedback while you are filling out custom forms for whether you are entering the correct data.

To test create a form, and add a form input that has is_required set to true. With the next commit you can also choose the validation type from a drop down in the admin to add to the is_required or use without the is_required to validate optional fields as well.