All Settings Must be Defined Now in the Setting Model.
This is a usability upgrade which makes adding settings for sites easier, because the available settings are pre-defined in the settings model.  IMPORTANT NOTE : Moving forward if you add a new setting to be available, you must put it into the settings model, so that it can be added at /admin/settings/add.

READ THAT NOTE AGAIN, TO MAKE SURE YOU UNDERSTAND

Also, we should look at the currently defined settings and think about how they can be better condensed into less settings.  Any settings which have to be set at the same time belong in a single setting, not a bunch of different settings.  Take the __ORDERS_SHIP_FROM_STATE,  __ORDERS_SHIP_FROM_ZIP for example.  That should not be two different settings, and should instead be a single constant with sub setting values under it.  Moving forward please do better grouping of settings.