Reusable Select Ajax Javascript.
You can now use some custom options in your select inputs in order to get that select input to update other select inputs.   Please look at the html source of /admin/settings/add, specifically for the Setting Type drop down to see how it works.

within the html itself, if the select has the attributes, [json], [element], and [rel], then that select will make an ajax call to the json url, and update the id of the value of rel with new select options.