Per Template CSS, and JS Upgrade.
You can now specify which template a css file or js file is for when uploading the code at /webpages/webpage_csses/add or edit and /webpages/webpage_jses/add or edit.  If you don't specify a specific template then it will be used for all templates, otherwise it will only be used for the exact template id that it is specified for.  (so if you need it for two different templates and not a third, then you would need to upload the css or js file using two different names.)

To test create a template, then upload css choosing that template from the drop down.