Added New AuthError Setting
Created a new setting so that each site could customize the message that displays when visitors reach a page which is restricted and they are not logged in.

Test it by going to /admin/settings/add and choose App from the drop down type, and follow the example message text.