Updated Admin Look & Feel
Its not done, but its worth checking out.  I have created a design that can be reused throughout the entire admin for nearly every system I can think of.  This was design with hasMany, belongsTo in mind, and should cover all of the data relationship cases that come up.   Therefore we should have a finished design pending any feedback.  NOTE : There is still more coding to do, but the design is done.