Added System Check for Settings Synchornization.
Now if defaults.ini and settings.ini do not match exactly then there is an error message printed on top of the site, so that you know that the site needs an upgrade.  This works because defaults.ini is tracked in this repo, and settings.ini files aren't.  So if someone makes local changes for example, then commits the updated defaults.ini file, anyone else downloads that repo will get a message that their local settings don't match the latest defaults.

This is good because you will immediately know when you visit a site whether it is properly setup for review, with the right settings that have been tested locally by someone else.

Though it should be said, that this is not the end of database synchronization problems, and if anyone has a good no hassle way to manage database synchronization, I would love to hear it.