Cookies for Helper Text and Font Size
When you change the font size [in the footer] or turn off helper text, it saves a cookie so that when you return it is the same.

IMPORTANT NOTE :  The settings are system wide if you do them from the /admin url, but if you do the change to any particular page, then it is particular to only that page.  So if you remove helper text at /admin/projects, it should leave it on for all other pages, etc.