Added Sitemaps Plugin.
This plugin connects to acl, and outputs a sitemap for all pages which have guest access.  Requires that all models do use the displayField variable in order to work right. To view the sitemap, go to /sitemap or if you need xml, go to /sitemap.xml.

It is also supposed to submit the sitemap to search engines, but I have not tested that quite yet.  That is next up.