Menus Made Safe for Non-Upgraded Sites.
Added a check to see if the menus table exists before making the Webpage model have an association to it.  If you don't do this, then sites which aren't upgraded will be broken until they upgrade.  I prefer that we deprecate and remove things checks like this gradually in order to give all sites a chance to upgrade first.