Converted CKEditor to Jquery Loading.
You now load ckeditor using jquery code, and that made it possible to hide the buttons by default.  This may also be the forerunner to having buttons on Ckeditor available in a db configurable way, as it will be easier to hide them in the CKE helper using site settings.