Added Setting Text for Already Existing __APP_REGISTRATION_EMAIL_VERIFICATION
Do not confuse this with the new setting being worked on now, for manual verification of user registration.   This is the setting for email verification.  With the note that I changed the name of the constant to reflect the difference. Previously it was __APP_USERS_REQUIRE_VERIFICATION which was too close to what the new setting should be named.