Added a little security to settings.
This isn't meant for hard core security or anything, but it will at least serve to keep passwords in the defaults.ini file out of github and any other public place zuha files end up as plain text.  And since the salt is stored in database.php it is ignored by our git hub repos.  And since you're not supposed to be able to decrypt without the salt, it should be a semi-secure way to store passwords in the ini files.