SMTP Settings Added.
You now have to set SMTP settings in order to send mail via Swiftmailer.  No longer does it automatically use our testing email address.  __SYSTEM_SMTP is the setting that has to be set.  Look at the description when choosing that setting in order to see the options which are required.