Cleaning up and Adding @todo messages.
Took out some code we don't need anymore, moved some around to be saved for later, and added comments in some spots which denote things that need to be done in the future.  Every chance you get to leave @todo messages is good, because we can build task lists from that later.