Workflows are a little easier now.
Did some work on making workflows easier to create.  The condition is now automatically created from the workflow form, instead of having to create a condition first, and then create a related workflow after wards.

You can create a simple workflow (ie. when Project Issue is created as a starting point) with about 3 clicks now.   Go to /workflows/workflows/add and fill out the now easy to understand form.  (sub conditions, are still not intuitive and easy, but had to start somewhere.)

You may also notice that this is the first plugin where we have started to get rid of the admin all together.  It is no longer necessary to use the admin_ prefix anywhere, anytime, and these can be deleted going forward as long as we save the functionality.

If we need a view file different for an admin, then we would create a folder within the views/controller folder called [administrators], and add separate view files there. And because the teamplates are driven by user group we still can have an admin look and feel without needing to have multiple version of view files which are essentially the same.