Catalog Item Entry is Easier Now.
Made the default layout more intuitive, by restructuring, adding better language, and hiding optional areas.

Also added a time picker in addition to the date picker, and made it part of the form helper, so that anytime there is a date field type or a datetime field type, they will use the datepicker instead of the multiple select boxes for year month date, hours minutes seconds.