Catalog Item Galleries by Attribute.
Now when you select an attribute, it loads the related gallery.  So if you have a red dress and a black dress, and different pictures for each one, then you will see black pictures when looking at the black attribute, and red when you have selected the red attribute.  Its really cool.