Categories Usability Updates.
When you visit the categories add page, you now choose what model this category is going to be for first.  This allows us to redirect to the categorized page afterwards and you can the assign your new category to any existing records in that model.  Previous to this you had to search for this categorized link and try to figure out why your category wasn't showing up in your list of categories to assign when adding something [like a catalog item].